Below set of changes are targeted to achieve the following
* Better organization of test suites and utilities for easier use
* Ability to run tests in parallel as well as in serial mode
* Enhanced reporting and logging
* Results to be stored in result.xml in Junit Test report format
to be later readable by tools like Jenkins

More details on running these tests can be found at
https://github.com/Juniper/contrail-test/wiki/Running-tests-in-parallel
Another example :
https://github.com/Juniper/contrail-test/wiki/Running-Neutron-API-Tests

+++ fixtures/project_test.py
@@ -1,3 +1,4 @@
+import os
 import fixtures
 from keystoneclient.v2_0 import client as ksclient
 from vnc_api.vnc_api import *
@@ -7,17 +8,20 @@ import fixtures
 from quantum_test import *
 from vnc_api_test import *
 from contrail_fixtures import *
-from connections import ContrailConnections
-from util import retry
+from common.connections import ContrailConnections
+from tcutils.util import retry
 from time import sleep
 from keystoneclient import exceptions as ks_exceptions
-from util import get_dashed_uuid
+from tcutils.util import get_dashed_uuid
 
 
 class ProjectFixture(fixtures.Fixture):
 
-    def __init__(self, vnc_lib_h, connections, project_name='admin', username=None, password=None, role='admin'):
+    def __init__(self, vnc_lib_h, connections, project_name=None,
+                 username=None, password=None, role='admin', scale= False):
         self.inputs = connections.inputs
+        if not project_name:
+            project_name = self.inputs.stack_tenant
         self.vnc_lib_h = vnc_lib_h
         self.connections = connections
         self.project_name = project_name
@@ -32,18 +36,23 @@ class ProjectFixture(fixtures.Fixture):
         self.tenant_dict = {}
         self.user_dict = {}
         self._create_user_set = {}
+        insecure = bool(os.getenv('OS_INSECURE',True))
         if not self.inputs.ha_setup:
-            self.auth_url = 'http://%s:5000/v2.0' % (self.inputs.openstack_ip)
+            self.auth_url = os.getenv('OS_AUTH_URL') or \
+                'http://%s:5000/v2.0' % (self.inputs.openstack_ip)
         else:
-            self.auth_url = 'http://%s:5000/v2.0' % (self.inputs.keystone_ip)
+            self.auth_url = os.getenv('OS_AUTH_URL') or \
+                'http://%s:5000/v2.0' % (self.inputs.openstack_ip)
         self.kc = ksclient.Client(
             username=self.inputs.stack_user,
             password=self.inputs.stack_password,
             tenant_name=self.inputs.project_name,
-            auth_url=self.auth_url)
+            auth_url=self.auth_url,
+            insecure=insecure)
         self.project_connections = None
         self.api_server_inspects = self.connections.api_server_inspects
         self.verify_is_run = False
+        self.scale = scale
     # end __init__
 
     def _create_project(self):
@@ -62,13 +71,25 @@ class ProjectFixture(fixtures.Fixture):
     # end _delete_project
 
     def _create_project_keystone(self):
-        if self.project_name == 'admin':
-            self.logger.info('Project admin already exist, no need to create')
+        if self.project_name == self.inputs.stack_tenant:
+            try:
+                ks_project = self.kc.tenants.find(name=self.project_name)
+                if ks_project:
+                    self.already_present = True
+                    self.project_id = get_dashed_uuid(ks_project.id)
+                    self.logger.debug(
+                        'Project %s already present.Not creating it' %
+                        self.project_fq_name)
+            except ks_exceptions.NotFound, e:
+                self.logger.info('Project %s not found' % (
+                    self.project_name))
+                raise e
+            self.project_obj = self.vnc_lib_h.project_read(id=self.project_id)
+            self.uuid = self.project_id
             return self
-        project_list_in_api_before_test = self.vnc_lib_h.projects_list()
-        print "project list before test: %s" % project_list_in_api_before_test
 
         # create project using keystone
+        self.logger.info('Proceed with creation of new project.')
         self.ks_project_id = self.kc.tenants.create(self.project_name).id
         self.logger.info('Created Project:%s, ID : %s ' % (self.project_name,
                                                            self.ks_project_id))
@@ -79,7 +100,13 @@ class ProjectFixture(fixtures.Fixture):
 
     def _delete_project_keystone(self):
         self.logger.info('Deleting Project %s' % self.project_fq_name)
-        self.kc.tenants.delete(self.tenant_dict[self.project_name])
+        try:
+            self.kc.tenants.delete(self.tenant_dict[self.project_name])
+        except ks_exceptions.ClientException, e:
+            # TODO Remove this workaround 
+            if 'Unable to add token to revocation list' in str(e):
+                self.logger.warn('Exception %s while deleting project' % (
+                    str(e)))
     # end _delete_project
 
     def _reauthenticate_keystone(self):
@@ -92,22 +119,28 @@ class ProjectFixture(fixtures.Fixture):
 
     def setUp(self):
         super(ProjectFixture, self).setUp()
-        try:
-            ks_project = self.kc.tenants.find(name=self.project_name)
-            if ks_project:
-                self.already_present = True
-                self.project_id = get_dashed_uuid(ks_project.id)
-                self.logger.debug(
-                    'Project %s already present.Not creating it' %
-                    self.project_fq_name)
-        except ks_exceptions.NotFound, e:
-            self.logger.info('Project %s not found, creating it' % (
-                self.project_name))
+        if self.scale:
             self._create_project_keystone()
-            time.sleep(2)
+        else:
+            try:
+                ks_project = self.kc.tenants.find(name=self.project_name)
+                if ks_project:
+                    self.already_present = True
+                    self.project_id = get_dashed_uuid(ks_project.id)
+                    self.logger.debug(
+                        'Project %s already present.Not creating it' %
+                        self.project_fq_name)
+            except ks_exceptions.NotFound, e:
+                self.logger.info('Project %s not found, creating it' % (
+                    self.project_name))
+                self._create_project_keystone()
+                time.sleep(2)
         self.project_obj = self.vnc_lib_h.project_read(id=self.project_id)
         self.uuid = self.project_id
     # end setUp
+
+    def getObj(self):
+        return self.project_obj
 
     def cleanUp(self):
         super(ProjectFixture, self).cleanUp()
@@ -119,6 +152,10 @@ class ProjectFixture(fixtures.Fixture):
         if self.inputs.fixture_cleanup == 'force':
             do_cleanup = True
         if do_cleanup:
+            if not self.check_no_project_references():
+                self.logger.warn('One or more references still present' 
+                    ', will not delete the project %s' % (self.project_name))
+                return
             self._reauthenticate_keystone()
             self._delete_project_keystone()
             if self.verify_is_run:
@@ -129,19 +166,53 @@ class ProjectFixture(fixtures.Fixture):
 
     # end cleanUp
 
+    @retry(delay=2, tries=10)
+    def check_no_project_references(self):
+        vnc_project_obj = self.vnc_lib_h.project_read(id=self.project_id)
+        vns = vnc_project_obj.get_virtual_networks()
+        if vns:
+            self.logger.warn('Project %s still has VNs %s before deletion' %(
+                self.project_name, vns))
+            return False
+        vmis = vnc_project_obj.get_virtual_machine_interfaces()
+        if vmis:
+            self.logger.warn('Project %s still has VMIs %s before deletion' %(
+                self.project_name, vmis))
+            return False
+        sgs = vnc_project_obj.get_security_groups()
+        if len(sgs) > 1:
+            self.logger.warn('Project %s still has SGs %s before deletion' %(
+                self.project_name, sgs))
+            return False
+        return True
+    # end check_no_project_references
+
     def get_from_api_server(self):
         self.project_obj = self.vnc_lib_h.project_read(
             fq_name=self.project_fq_name)
         return self.project_obj
 
+    @retry(delay=2, tries=10)
+    def check_for_VN_in_api(self):
+        self.project_obj = self.vnc_lib_h.project_read(
+	    fq_name=self.project_fq_name)
+        has_vns = self.project_obj.get_virtual_networks()
+        if has_vns:
+            self.logger.info("Following VNs exist in project: %s" %has_vns)
+            return False
+        else:
+            self.logger.info("Don't see any VNs in the project %s" %self.project_fq_name)
+            return True
+
     def get_project_connections(self, username=None, password=None):
         if not username:
-            username = self.username or 'admin'
+            username = self.username or self.inputs.stack_user
         if not password:
-            password = self.password or 'contrail123'
+            password = self.password or self.inputs.stack_password
         if not self.project_connections:
             self.project_connections = ContrailConnections(
                 inputs=self.inputs,
+                logger=self.logger,
                 project_name=self.project_name,
                 username=username,
                 password=password)
@@ -171,7 +242,7 @@ class ProjectFixture(fixtures.Fixture):
                 result &= False
                 return result
             if cs_project_obj['project']['uuid'] != self.project_id:
-                self.logger.warn('Project id %s got from API Server %s'
+                self.logger.warn('Project id %s got from API Server is'
                                  ' not matching expected ID %s' % (
                                      cs_project_obj['project']['uuid'], self. project_id))
                 result &= False
@@ -188,6 +259,8 @@ class ProjectFixture(fixtures.Fixture):
             cs_project_obj = api_s_inspect.get_cs_project(
                 self.domain_name,
                 self.project_name)
+            self.logger.info("Check for project %s after deletion, got cs_project_obj %s" %
+                (self.project_name, cs_project_obj))
             if cs_project_obj:
                 self.logger.warn('Project %s is still found in API Server %s'
                                  'with ID %s ' % (self.project_name, api_s_inspect._ip,
